4.14.4 Update
Fixed an oversight in the fmin_cg_minimize routine: the functions from
Krauth and Lichtenstein consider as maxfn not the number of iterations
but the number of calls to the function to minimize. This includes the
calls for the numerical gradient.

Fixed by rescaling the number of iterations to
2*len(parameter_vector)*itermax

Also added message if the function returns an iexit value != 1