Intermediate commit
Starting documentation. One function per day, coffee break activity.
It gives the chance of cleaning up some code as well.
For example, the function "linspace" was defined in three places and, as
far as I can see, did the same thing. It is probably better to have it
defined in a single place and include it from the module where
necessary.