Intermediate commit:
in parpack, when the block size is reset a warning message is printed.
However since the block size is reset quite often especially if many
cores are used, and this message is printed by all cores, the logfile is
flooded.

This is just a warning, and diagonalization is still fine if it is
triggere. So put the message behind a verbose flag, and let only the
master print it