Modify (p)arpack fail logic
As of now, if info /= 0 the postprocessing routine is never called.
This results in a situation where spectra are not calculated at all
but the code doesn't stop. In e.g. a Fock space sector solution, this
results in some sectors having all 0 eigenvalues.

If info >0 this is a soft fail. Usually, the postprocessing routine can
still be called but the spectrum will not be precise. It is probably
better than nothing and certainly better than having all zeros.

This commit prints a message in case of soft fail, but still calls the
postprocessing routine. If info < 0 that is a hard fail, and the program
stops.